# docs/api/openapi.yaml
...
openapi: 3.1.0
...
info:
    title: Footnote Backend API
    # Versioning: bump when request/response shapes, status codes, or auth requirements change; docs-only edits can skip.
    version: 1.6.0
    license:
        name: MIT + Hippocratic License v3 (HL3-CORE)
        url: docs/LICENSE_STRATEGY.md
    description: |
        Backend wire contract and integration boundary for other packages.
        Chat is the shared backend action-routing boundary for web and Discord surfaces.
        Non-2xx responses use ErrorResponse unless an operation states otherwise.
        info.version is manually maintained (see docs/api/README.md).

# Explicitly mark endpoints as unauthenticated by default.
security: []
...
servers:
    - url: https://{backendHost}
      description: Backend API base URL
      variables:
          backendHost:
              default: api.your-domain.tld

# ====================
# Tags
# ====================
...
paths:
...
    /api/incidents/report:
        post:
            operationId: postIncidentReport
            x-codeRefs:
                - packages/backend/src/handlers/incidents.ts#handleIncidentReportRequest
                - packages/api-client/src/incidents.ts#reportIncident
                - packages/contracts/src/web/types.ts#PostIncidentReportRequest
                - packages/contracts/src/web/types.ts#PostIncidentReportResponse
                - packages/contracts/src/web/schemas.ts#PostIncidentReportResponseSchema
            summary: Create an incident report
            description: |
                Trusted internal endpoint for Discord-driven incident reports. Returns a one-time, 30-day claim code to the trusted bot; only its hash is persisted, and the code is shown privately to the reporter for a later explicit association.
                Accepts raw Discord resource identifiers from the bot and persists only pseudonymized pointers.
                Requires either `x-service-token` or `x-trace-token`.
                Intended for trusted internal callers; normal internal service rate limits still apply.
            tags:
                - Incidents
            security:
                - TrustedServiceToken: []
                - TraceToken: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/IncidentReportRequest'
                        example:
                            reporterUserId: '123456789012345678'
                            guildId: '234567890123456789'
                            channelId: '345678901234567890'
                            messageId: '456789012345678901'
                            jumpUrl: 'https://discord.com/channels/234567890123456789/345678901234567890/456789012345678901'
                            responseId: 'response_123'
                            chainHash: 'hash_abc'
                            modelVersion: 'gpt-5-mini'
                            tags: [safety, review]
                            description: 'The assistant gave risky instructions.'
                            contact: 'moderator@example.com'
                            consentedAt: '2026-03-13T17:00:00.000Z'
            responses:
                '200':
                    $ref: '#/components/responses/IncidentReportOk'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '405':
                    $ref: '#/components/responses/MethodNotAllowed'
                '413':
                    $ref: '#/components/responses/PayloadTooLarge'
                '500':
                    $ref: '#/components/responses/InternalServerError'
    /api/incidents:
        get:
            operationId: listIncidents
            x-codeRefs:
                - packages/backend/src/handlers/incidents.ts#handleIncidentListRequest
                - packages/api-client/src/incidents.ts#listIncidents
                - packages/contracts/src/web/types.ts#GetIncidentsResponse
            summary: List incidents
            description: |
                Trusted internal endpoint for newest-first operator-safe incident summaries.
                Requires either `x-service-token` or `x-trace-token`.
                Intended for trusted internal callers; normal internal service rate limits still apply.
            tags:
                - Incidents
            security:
                - TrustedServiceToken: []
                - TraceToken: []
            parameters:
                - name: status
                  in: query
                  required: false
                  schema:
                      $ref: '#/components/schemas/IncidentStatus'
                - name: tag
                  in: query
                  required: false
                  schema:
                      type: string
                - name: createdFrom
                  in: query
                  required: false
                  schema:
                      type: string
                      format: date-time
                - name: createdTo
                  in: query
                  required: false
                  schema:
                      type: string
                      format: date-time
            responses:
                '200':
                    $ref: '#/components/responses/IncidentListOk'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '405':
                    $ref: '#/components/responses/MethodNotAllowed'
                '500':
                    $ref: '#/components/responses/InternalServerError'
    /api/incidents/{incidentId}:
        get:
            operationId: getIncident
            x-codeRefs:
                - packages/backend/src/handlers/incidents.ts#handleIncidentDetailRequest
                - packages/api-client/src/incidents.ts#getIncident
                - packages/contracts/src/web/types.ts#GetIncidentResponse
            summary: Fetch one incident
            description: |
                Trusted internal endpoint for one operator-safe incident detail payload.
                Requires either `x-service-token` or `x-trace-token`.
                Intended for trusted internal callers; normal internal service rate limits still apply.
            tags:
                - Incidents
            security:
                - TrustedServiceToken: []
                - TraceToken: []
            parameters:
                - $ref: '#/components/parameters/IncidentId'
            responses:
                '200':
                    $ref: '#/components/responses/IncidentDetailOk'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '405':
                    $ref: '#/components/responses/MethodNotAllowed'
                '500':
                    $ref: '#/components/responses/InternalServerError'
    /api/incidents/{incidentId}/status:
        post:
            operationId: postIncidentStatus
            x-codeRefs:
                - packages/backend/src/handlers/incidents.ts#handleIncidentStatusRequest
                - packages/api-client/src/incidents.ts#updateIncidentStatus
                - packages/contracts/src/web/types.ts#PostIncidentStatusRequest
                - packages/contracts/src/web/types.ts#PostIncidentStatusResponse
            summary: Update incident status
            description: |
                Trusted internal endpoint for operator-driven incident status changes.
                Requires either `x-service-token` or `x-trace-token`.
                Intended for trusted internal callers; normal internal service rate limits still apply.
            tags:
                - Incidents
            security:
                - TrustedServiceToken: []
                - TraceToken: []
            parameters:
                - $ref: '#/components/parameters/IncidentId'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/IncidentStatusRequest'
                        example:
                            status: under_review
                            actorUserId: '999999999999999999'
                            notes: 'Reviewing report context now.'
            responses:
                '200':
                    $ref: '#/components/responses/IncidentDetailOk'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '405':
                    $ref: '#/components/responses/MethodNotAllowed'
                '413':
                    $ref: '#/components/responses/PayloadTooLarge'
                '500':
                    $ref: '#/components/responses/InternalServerError'
    /api/incidents/{incidentId}/notes:
        post:
            operationId: postIncidentNotes
            x-codeRefs:
                - packages/backend/src/handlers/incidents.ts#handleIncidentNotesRequest
                - packages/api-client/src/incidents.ts#addIncidentNote
                - packages/contracts/src/web/types.ts#PostIncidentNotesRequest
                - packages/contracts/src/web/types.ts#PostIncidentNotesResponse
            summary: Append an incident note
            description: |
                Trusted internal endpoint for adding operator notes without changing incident status.
                Requires either `x-service-token` or `x-trace-token`.
                Intended for trusted internal callers; normal internal service rate limits still apply.
            tags:
                - Incidents
            security:
                - TrustedServiceToken: []
                - TraceToken: []
            parameters:
                - $ref: '#/components/parameters/IncidentId'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/IncidentNotesRequest'
                        example:
                            actorUserId: '999999999999999999'
                            notes: 'Escalating to the safety review queue.'
            responses:
                '200':
                    $ref: '#/components/responses/IncidentDetailOk'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '405':
                    $ref: '#/components/responses/MethodNotAllowed'
                '413':
                    $ref: '#/components/responses/PayloadTooLarge'
                '500':
                    $ref: '#/components/responses/InternalServerError'
    /api/incidents/{incidentId}/remediation:
        post:
            operationId: postIncidentRemediation
            x-codeRefs:
                - packages/backend/src/handlers/incidents.ts#handleIncidentRemediationRequest
                - packages/api-client/src/incidents.ts#recordIncidentRemediation
                - packages/contracts/src/web/types.ts#PostIncidentRemediationRequest
                - packages/contracts/src/web/types.ts#PostIncidentRemediationResponse
            summary: Record incident remediation outcome
            description: |
                Trusted internal endpoint for recording the bot's immediate remediation attempt.
                Requires either `x-service-token` or `x-trace-token`.
                Intended for trusted internal callers; normal internal service rate limits still apply.
            tags:
                - Incidents
            security:
                - TrustedServiceToken: []
                - TraceToken: []
            parameters:
                - $ref: '#/components/parameters/IncidentId'
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/IncidentRemediationRequest'
                        example:
                            actorUserId: '999999999999999999'
                            state: applied
                            notes: 'Warning banner applied with spoiler wrapping.'
            responses:
                '200':
                    $ref: '#/components/responses/IncidentDetailOk'
                '400':
                    $ref: '#/components/responses/BadRequest'
                '401':
                    $ref: '#/components/responses/Unauthorized'
                '403':
                    $ref: '#/components/responses/Forbidden'
                '404':
                    $ref: '#/components/responses/NotFound'
                '405':
                    $ref: '#/components/responses/MethodNotAllowed'
                '413':
                    $ref: '#/components/responses/PayloadTooLarge'
                '500':
                    $ref: '#/components/responses/InternalServerError'

    # Trace upsert (TraceToken required, rate-limited).
...
components:
...
    securitySchemes:
        # Shared secret for trace ingestion.
        TraceToken:
            type: apiKey
            in: header
            name: x-trace-token
            description: Shared secret for trusted trace ingestion.
        TrustedServiceToken:
            type: apiKey
            in: header
            name: x-service-token
            description: Shared secret for trusted internal service callers.
        AgentToken:
            type: apiKey
            in: header
            name: x-agent-token
            description: Dedicated trusted credential for agent chat callers.
        AdminSettingsToken:
            type: apiKey
            in: header
            name: x-admin-token
            description: Dedicated trusted admin token for settings endpoints.
        SetupSessionCookie:
            type: apiKey
            in: cookie
            name: footnote_setup_session
            description: Short-lived setup session cookie accepted while the session remains valid.
        AccountSessionCookie:
            type: apiKey
            in: cookie
            name: footnote_account_session
            description: Opaque local account session cookie. It contains no identity claims or provider tokens.
        DiscordConnectionCookie:
            type: apiKey
            in: cookie
            name: footnote_discord_connection
            description: Opaque, short-lived browser binding for one Discord connection transaction.

    # Data models for request/response payloads.
    schemas:
...
        IncidentReportRequest:
            type: object
            properties:
                reporterUserId:
                    type: string
                guildId:
                    type: string
                channelId:
                    type: string
                messageId:
                    type: string
                jumpUrl:
                    type: string
                    format: uri
                responseId:
                    type: string
                chainHash:
                    type: string
                modelVersion:
                    type: string
                tags:
                    type: array
                    items:
                        type: string
                description:
                    type: string
                contact:
                    type: string
                consentedAt:
                    type: string
                    format: date-time
            required:
                - reporterUserId
                - consentedAt
            additionalProperties: false
            example:
                reporterUserId: '123456789012345678'
                guildId: '234567890123456789'
                channelId: '345678901234567890'
                messageId: '456789012345678901'
                jumpUrl: 'https://discord.com/channels/234567890123456789/345678901234567890/456789012345678901'
                responseId: 'response_123'
                chainHash: 'hash_abc'
                modelVersion: 'gpt-5-mini'
                tags: [safety, review]
                description: 'The assistant gave risky instructions.'
                contact: 'moderator@example.com'
                consentedAt: '2026-03-13T17:00:00.000Z'
...